Feature/R4SU-55 game code refactoring
