+++ docs/game.md
@@ -44,39 +44,53 @@ _Можно добавить:_
 
 ### Архитектура игры
 
+gameEngine/core: - основная логика игры
+
+    baseObject.ts - базовый обьект игры для отрисовки, у которого есть общий тип состояния с координатами для отрисовки и общий тип параметров с размерами.
+
+    collisionModule.ts - определяет коллизии по текущему стейту и апдейтает стейт при столкновении
+
+    controlModule.ts - модуль отвечающий за управление игрой
+
+    engine.ts - отвечает за главный цикл анимации игры, его запуск, сброс и остановка. В цикле проверяется актуальность обьектов умер/жив корабль, виден/спрятан выстрел, умер ли плейер, а также закончился ли уровень. Происходит апдейт стейта всех обьектов (кораблей, выстрелов). Вызывается painter для отрисовки нужного обьекта. Проверяются коллизии. Выставляется нужный стейт игры в глобальном хранилище, в зависимости от глобального стейта загружается нужный экран (обложки игры до старта, завершение игры, завершение уровня, конец игры, старт, пауза)
+
+    painter.ts - умеет отрисовывать базовые обьекты с картинкой, координатами и размерами, умеет рисовать бекграунд, экран начала игры и экран
+
 gameEngine/helpers: - вспомогательные методы рассчетов связанные с игрой
 
-trajectoryHelper.ts - вспомогательные методы для рассчета траектории полета
+    trajectoryHelper.ts - вспомогательные методы для рассчета траектории полета
 
-gameEngine/parameters (все параметры которые задаются один раз на всю игру):
+gameEngine/objects:
 
-gameLevels.ts - вся логика связанная с уровнями игры, сколько уровней, описание логики каждого уровня: время уровня, бонусы на уровне, количество врагов каждого типа и траектория их движения/выстрелы, особые ловушки врагов и т.п.
+    base:
 
-gameParameters.ts - общие параметры игры: размер игровой области, фоновое изображение и т.п.
+            ship.ts - базовый тип корабля
 
-gameObjectsParameters.ts - параметры игровых обьектов: размеры и картинки для выстрелов, размеры и картинки для врагов и игрока
+            shot.ts - базовый тип выстрела
 
-gameEngine/store:
+    enemyShip.ts - общий тип кораблей врага
 
-mockGameState.ts - моковый прототип стейта игры (впоследствии стейт будет хранится в редакс). Стейт хранит состояние игры: текущий уровень, массив кораблей врагов, корабль игрока, массив выстрелов, бонусов для отрисовки.
-
-gameEngine/types:
-
-    commonTypes.ts - общие типы игры: точка координат, тип корабля, тип выстрела, а также общий тип отрисовываемого обьекта (DrawableGameObject) у которого есть общий тип состояния с координатами для отрисовки и общий тип параметров с размерами.
-
-    gameTypes.ts - игровые типы: общий тип отрисовываемого обьекта, его наследники корабль и выстрел
-
-    objectState.ts - все типы относящиеся к стейтам игровых обьектов (координаты, стейт жизненного цикла и т.д.)
+    playerShip.ts - корабль игрока
 
     trajectory.ts - класс отвечающий за плавную траекторию движения корабля, может вычислять путь корабля по заданному внутри уровня массиву точек. Экземпляр траектории хранится в каждом экземпляре вражеского корабля чтобы он знал как ему лететь на данном уровне (логика инкапсулирована)
 
-gameEngine:
+gameEngine/parameters (все параметры которые задаются один раз на всю игру):
 
-    collisionManager.ts - определяет коллизии по текущему стейту и выставляет стейт при столкновении (todo апдейт стейта вынести?)
+    levels.ts - вся логика связанная с уровнями игры, сколько уровней, описание логики каждого уровня: время уровня, бонусы на уровне, количество врагов каждого типа и траектория их движения/выстрелы, особые ловушки врагов и т.п.
 
-    gameEngine.ts - отвечает за главный цикл анимации игры, его запуск, сброс и остановка. В цикле проверяется актуальность обьектов умер/жив корабль, виден/спрятан выстрел, умер ли плейер, а также закончился ли уровень. Происходит апдейт стейта всех обьектов (кораблей, выстрелов). Вызывается gamePainter для отрисовки нужного обьекта. Проверяются коллизии. Выставляется нужный стейт игры в глобальном хранилище, в зависимости от глобального стейта загружается нужный экран (обложки игры до старта, завершение игры, завершение уровня, конец игры, старт, пауза)
+    globalParameters.ts - общие параметры игры: размер игровой области, фоновое изображение и т.п.
 
-    gamePainter.ts - умеет отрисовывать базовые обьекты с картинкой, координатами и размерами, умеет рисовать бекграунд, экран начала игры и экран
+    objectsParameters.ts - параметры игровых обьектов: размеры и картинки для выстрелов, размеры и картинки для врагов и игрока
+
+gameEngine/store:
+
+    gameState.ts - cтейт игры (планирую перенести стейт в редакс). Стейт хранит состояние игры: текущий уровень, массив кораблей врагов, корабль игрока, массив выстрелов, бонусов для отрисовки.
+
+    objectState.ts - стейты обьектов игры, хранят координаты, знают на каком этапе жизненного цикла корабль (ожидает вылета, летит, взрывается, умер).
+
+gameEngine/types:
+
+    common.ts - общие типы игры: точка координат, тип корабля, тип выстрела,
 
 game.module.scss - стили для компонента игры
 game.tsx - компонент игры
