+++ packages/client/src/app/pages/game/game.tsx
@@ -2,9 +2,9 @@ import React, { FC, useEffect, useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
 import style from './game.module.scss';
 import Button from '@/app/components/common/button/button';
-import params from './gameEngine/parameters/gameParameters';
-import GameEngine from './gameEngine/gameEngine';
-import { GlobalGameState } from './gameEngine/types/objectState';
+import params from './gameEngine/parameters/globalParameters';
+import GameEngine from './gameEngine/core/engine';
+import { GlobalGameState } from './gameEngine/store/objectState';
 import { RootState } from '@/app/store/store';
 import GameOver from '@/app/components/gameOver/gameOver';
 import StartGame from '../startGame/startGame';
