+++ packages/client/src/app/pages/game/gameEngine/parameters/globalParameters.ts
